Fix #17300: now captions() is obsolete in favor of legend()
As `legend` called the internal `caption` (sci_Legend() gateway), whose features are a subset of `legend` features, we propose that 
1. cpp gateway `sci_Legend()` is renamed `sci_percent_legend()` (`%_legend` at the interpreter level)
2. like `legend`, `caption` becomes a macro calling `%_legend`, but reverses the order of the elements of the handle vector
3. all internals calling captions() now call `%_legend()` instead
4. `captions` is marked as obsolete

The code in relevant C++ and Java functions has been simplified by removing the numerous back and forth changes of order (which are useless now) for handle and/or string arguments.

Remark: now when a legend entity is created, its `links` property is now a vector with the same order as the legend strings.

Co-authored-by: Antoine ELIAS <antoine.elias@3ds.com>
Co-authored-by: Vincent COUVERT <vincent.couvert@3ds.com>
Closes #17300
See https://gitlab.com/scilab/scilab/-/merge_requests/1059