unifying graphic commands output after c1dbf528
Since c1dbf528, Scilab gateways and macros know when no output argument is given. It has already allowed to improve the API of numerous macros and gateways. In this commit, among graphic functions, only plot3d(), surf() and plot() used the new API: when no output argument is given, nothing is returned and the handle or array of handles is returned when an output argument is given. This MR allows all relevant graphic functions (those who create graphic entities) to benefit from the new API. In most cases, the yielded output is the current entity (yielded by gce()), which allows to simplify and clarify the code. 
There are many cases where the current entity is a Compound with children sorted in the inverse order of creation, e.g. xfpolys. In that case, the output of xfpolys is a vector of handles sorted in natural order (this is the rule for Matlab graphic functions). 
There are other cases where functions where yielding a handle regardless of the number of output arguments, which was sometimes annoying (legend() and uicontrol()). In that case, the new API has been applied, which won't change anything for code using an explicit (other than "ans") output argument but will improve user experience (as before, this is the rule in Matlab for legend() and uicontrol()).

In all cases, the current entity yielded by gce() is unchanged, i.e. actual code using it will work as before.

Known incompatibilities:

- `fec` no more creates a `Compound` object.
- `h = plot(x1, y1, x2, y2)` now returns handles in `h` as in arguments i.e. `[h1, h2]`.

Co-authored-by: acarnis <adeline.carnis@3ds.com>
See https://gitlab.com/scilab/scilab/-/merge_requests/1022